requirements.txt
python-dotenv~=1.0.1
requests~=2.31.0
streamlit~=1.31.1